+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - Global variables can be given default string values using `global NAME = "STRING"` syntax.
+- Function `eq` to compare two values for equality.
 
 #### Removed